+++ src/pipes/index.ts
@@ -2,3 +2,4 @@ export * from './addNewOption/addNewOption.pipe';
 export * from './displayValue/displayValue.pipe';
 export * from './getPlugin/getPlugin.pipe';
 export * from './groupedListOptions/groupedListOptions.pipe';
+export * from './hasValue/hasValue.pipe';